requirements.txt
aniso8601==3.0.0
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
PyJWT==1.4.2
pytz==2018.3
six==1.11.0
Werkzeug==0.14.1
